+++ rse-cft-lib/src/main/java/uk/gov/hmcts/rse/ccd/lib/injected/Common.java
@@ -1,4 +1,8 @@
 package uk.gov.hmcts.rse.ccd.lib.injected;
 
+/**
+ * Marker interface for the package that is injected into
+ * each spring context.
+ */
 public interface Common {
 }
